Don't add newlines between archived headlines
I kept getting newlines between headlines I'm archiving, and I think
this is the cause.  We just moved to (point-max), so if (bolp) then
we're at a blank line--we don't need to insert another.  It's not only
datetree users that want this behavior, right?  Surely?

(Also, if we're worried about narrowing, shouldn't we just
(save-restriction (widen) ...) or something before we do our work in
this function?)