Improve fontification of nested emphasis
This fixes fontification of nested emphasis in two ways.  First, take
this example:

    *bold =pew *bang* pew= bold*

This was being fontified like:

    *bold      : bold
    =pew *bang*: bold and verbatim
    pew=       : verbatim
    bold*      : normal

Now I'm going to break a #emacs rule and quote someone there who told
me, "Parsing [of emphasis] is done from left to right, and emphasis
markers cannot overlap."  This is true *even when a marker is within
some code/verbatim emphasis*.  You can verify this by calling
`org-element-context` around the various parts of this line.  You will
find that `org-element-context` reports:

    *bold =pew *bang*: bold
    pew= bold*       : normal

This happens because the "*" after "bang" ends the bold emphasis by
the above rule of parsing left to right.  The possible existence of
verbatim emphasis starting at "=pew" is not even considered until
after the bold emphasis has been parsed as shown.  However, after that
bold emphasis has been parsed, the end of the "=pew" emphasis cannot
be found *within the existing bold emphasis*, so because overlapping
emphasis is not allowed, "=pew" does not start verbatim emphasis.

This commit makes the fontification match these results of
`org-element-context`.  This is accomplished by repurposing the
existing `org-emphasis` text property: heretofore this property was
merely t if the text was part of emphasis.  This property's value is
now `(VERBATIM-P . END)`, where `VERBATIM-P` is non-nil if the
emphasis is verbatim emphasis, and `END` is the position of the end of
the emphasis.  `END` is used to eliminate incorrect, overlapping
fontification.

`VERBATIM-P` is used to fix the second problem.  Here's an example
that illustrates this problem prior to this commit:

    =verb *pow* verb=

Prior to this commit, "*pow*" would be bold.  This is incorrect, as
markup within verbatim emphasis is ignored.  Once again,
`org-element-context` reports the above sample as being verbatim for
the entire line, no bold in the middle.  This commit uses the new
`VERBATIM-P` element of the `org-emphasis` property to avoid
fontifying things that look like emphasis inside code and verbatim
emphasis.

I'm hoping there is no problem in repurposing the `org-emphasis` text
property in this way.  The property was previously either absent or t,
and the only user of this property that I found within org itself only
tests whether or not the property is non-nil.